[refactor] Restructure Teams and Slack integration for improved organization
- Moved Slack and Teams specific imports to their respective modules under `requester` for better clarity and organization.
- Refactored `main.py` to utilize the new import paths, enhancing maintainability.
- Removed the now obsolete `requester_context.py` file, centralizing configuration management.
- Updated linting rules in `pyproject.toml` to reflect the new module structure.
- Cleaned up unused imports and optimized existing functions across various files.

This refactor streamlines the integration architecture, making it easier to manage and extend functionalities for both Slack and Teams.
