+++ src/requester/teams/teams_activity_helpers.py
